added util.js
