+++ tools/importer/transformers/index.js
@@ -1,8 +1,13 @@
 import header from './header.js';
 import footer from './footer.js';
 import stickyFooter from './stickyFooter.js';
+import videoEmbed from './videoEmbed.js';
 
 // eslint-disable-next-line import/prefer-default-export
+export const transformers = [
+  videoEmbed,
+];
+
 export const xfTransformers = [
   footer,
   stickyFooter,
